Made Link files script way better

+++ link_files.sh
@@ -2,28 +2,21 @@
 
 DOT_DIR=$(pwd)
 (
-rm -rf ~/.zshrc
-ln -s $DOT_DIR/.zshrc ~/
+for i in $DOT_DIR\/{.*rc,.*aliases,.vim}; do
+    [ -e "$i" ] || continue
+    i=${i##*/}
+    rm -rf ~/$i
+    ln -sf $DOT_DIR/$i ~/
+done
 
-rm -rf ~/.bashrc
-ln -s $DOT_DIR/.bashrc ~/
+if [[ -d ~/.zprezto ]]; then
+    rm -rf ~/.zprezto/modules/prompt/functions/prompt_timerty_setup
+    ln -sf $DOT_DIR/prompt_timerty_setup ~/.zprezto/modules/prompt/functions/
+fi
 
-rm -rf ~/.shrc
-ln -s $DOT_DIR/.shrc ~/
-
-rm -rf ~/.sh_aliases
-ln -s $DOT_DIR/.sh_aliases ~/
-
-rm -rf ~/.vim
-ln -s $DOT_DIR/.vim ~/
-
-rm -rf ~/.vimrc
-ln -s $DOT_DIR/.vimrc ~/
-
-rm -rf ~/.zprezto/modules/prompt/functions/prompt_timerty_setup
-ln -s $DOT_DIR/prompt_timerty_setup ~/.zprezto/modules/prompt/functions/
-
-rm -rf ~/.config/kglobalshortcutsrc
-ln -s $DOT_DIR/kglobalshortcutsrc ~/.config/
+if [[ -e ~/.config/kglobalshortcutsrc ]]; then
+    rm -rf ~/.config/kglobalshortcutsrc
+    ln -sf $DOT_DIR/kglobalshortcutsrc ~/.config/
+fi
 )
